+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # ChangeLog
 
+## 2.5.4
+機能追加
+ * `g.Object2D#anchorX` と `g.Object2D#anchorY` を追加
+
+### ゲーム開発者への影響
+
+ * `g.Object2D#anchorX` と `g.Object2D#anchorY` を追加
+    * ゲーム開発者は、これらのプロパティを指定することで、対象のオブジェクトのアンカーを設定することができます。
+
 ## 2.5.3
 
 機能追加